% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLossFunctionNormalOPC.R
\name{BayesianLossFunctionNormalOPC}
\alias{BayesianLossFunctionNormalOPC}
\title{BayesianLossFunctionNormalOPC: BayesianLossFunctionNormalOPC}
\usage{
BayesianLossFunctionNormalOPC(mu, sigma2, N, mu0, sigma02, N0, N0_max,
  weibull_scale, weibull_shape, number_mcmc, H0, two_side, inequality)
}
\arguments{
\item{mu}{numeric}

\item{sigma2}{numeric}

\item{N}{numeric}

\item{mu0}{numeric}

\item{sigma02}{numeric}

\item{N0}{numeric}

\item{N0_max}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{number_mcmc}{numeric}

\item{H0}{numeric}

\item{two_side}{character}

\item{inequality}{character}
}
\description{
BayesianLossFunctionNormalOPC
}
\details{
BayesianLossFunctionNormalOPC
}
\examples{

}

