% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLossFunctionSimulation.R
\name{BayesianLossFunctionSimulation}
\alias{BayesianLossFunctionSimulation}
\title{BayesianLossFunctionSimulation: BayesianLossFunctionSimulation}
\usage{
BayesianLossFunctionSimulation(y, N, y0, N0, N0_max, alpha0, beta0, number_mcmc,
  weibull_shape, weibull_scale, two_side)
}
\arguments{
\item{y}{numeric}

\item{N}{numeric}

\item{y0}{numeric}

\item{N0}{numeric}

\item{N0_max}{numeric}

\item{alpha0}{numeric}

\item{beta0}{numeric}

\item{number_mcmc}{numeric}

\item{weibull_shape}{numeric}

\item{weibull_scale}{numeric}

\item{two_side}{character}
}
\description{
BayesianLossFunctionSimulation
}
\details{
BayesianLossFunctionSimulation
}
\examples{

}

