% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial1arm.R
\name{bdpbinomial1arm}
\alias{bdpbinomial1arm}
\title{bdpbinomial1arm: bdpbinomial1arm}
\usage{
bdpbinomial1arm(y_t = 1, N_t = 400, y0_t = 10, N0_t = 100,
  alpha_max = 1, a0 = 1, b0 = 1, number_mcmc = 10000,
  weibull_scale = 0.05, weibull_shape = 2, H0 = 10, two_side = 0,
  inequality = "<")
}
\arguments{
\item{y_t}{numeric}

\item{N_t}{numeric}

\item{y0_t}{numeric}

\item{N0_t}{numeric}

\item{alpha_max}{numeric}

\item{a0}{numeric}

\item{b0}{numeric}

\item{number_mcmc}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{H0}{numeric}

\item{two_side}{character}

\item{inequality}{character}
}
\description{
bdpbinomial1arm
}
\details{
bdpbinomial1arm
}
\examples{

}

