% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial2arm.R
\name{bdpbinomial2arm}
\alias{bdpbinomial2arm}
\title{bdpbinomial2arm: bdpbinomial2arm}
\usage{
bdpbinomial2arm(y_t = 1, N_t = 400, y_c = 5, N_c = 400, y0_t = 1,
  N0_t = 200, y0_c = 5, N0_c = 200, alpha_max = 1, a0 = 1, b0 = 1,
  number_mcmc = 10000, weibull_scale = 0.05, weibull_shape = 2,
  two_side = 0, inequality = "<", delta = 0)
}
\arguments{
\item{y_t}{numeric}

\item{N_t}{numeric}

\item{y_c}{numeric}

\item{N_c}{numeric}

\item{y0_t}{numeric}

\item{N0_t}{numeric}

\item{y0_c}{numeric}

\item{N0_c}{numeric}

\item{alpha_max}{numeric}

\item{a0}{numeric}

\item{b0}{numeric}

\item{number_mcmc}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{two_side}{character}

\item{inequality}{character}

\item{delta}{character}
}
\description{
This code is used for estimating posterior samples from a binary outcome
where an informative prior is used. The prior weight is determined using a
loss function. In addition this code simulate many trials in order to get
trial characteristics you must specify the parameters of the loss function
as well as the maximum strength for the prior. This code assumes a
non-adaptive trial.
This code is modeled after the methodologies developed by the MDIC working
group: "Informing clinical trials using bench & simulations"
Section 1: of the code defines functions needed
Section 2: of the code estimates a posterior and the loss function value
           given inputs
Section 3: of the code simulates the trial many times to get trial
           characteristics
Developer: Tarek Haddad
Tarek.D.Haddad@Medtronic.com
Last modified:1/26/2016
}
\details{
bdpbinomial2arm
}
\examples{

}

