% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdplm.R
\name{bdplm}
\alias{bdplm}
\alias{bdplm,ANY-method}
\title{Bayesian Discount Prior: Two-Arm Linear Regression}
\usage{
bdplm(formula = formula, data = data, prior_mean = 0, prior_sigma = 1,
  number_mcmc_alpha = 10000, number_mcmc_sigma2grid = 1000,
  number_mcmc_sigma2 = 100, number_mcmc_beta = 10000, alpha_max = 1,
  fix_alpha = FALSE, number_mcmc = 10000, weibull_scale = 0.135,
  weibull_shape = 3, two_side = TRUE, method = "fixed")
}
\arguments{
\item{formula}{an object of class "formula." See "Details" for
more information, including specfication of historical and treatment
data status.}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)
containing the variables in the model. If not found in data,
the variables are taken from environment(formula), typically
the environment from which bdpregression is called.}

\item{prior_mean}{historical mean for the coefficients; default is 0.
Can be a vector of length equal to the number of covariates.
If it is a scalar, it is expanded to the length of this vector.}

\item{prior_sigma}{prior standard deviation for the coefficients:
default is 1. Can be a vector of length equal to the number of covariates.
If it is a scalar, it is expanded to the length of this vector.}

\item{number_mcmc_alpha}{scalar. Number of Monte Carlo
simulations for estimating the historical data weight. Default is 10000.}

\item{number_mcmc_sigma2grid}{scalar. Grid size for computing sigma2.
Default is 1000. See "Details" for more information.}

\item{number_mcmc_sigma2}{scalar. Number of Monte Carlo simulations for
estimating sigma2. Default is 1000. See "Details" for more information.}

\item{number_mcmc_beta}{scalar. Number of Monte Carlo simulations for
estimating beta, the vector of regression coefficients. Default is 10000.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. Users may specify a vector of two values where the first
value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}

\item{fix_alpha}{logical. Fix alpha at alpha_max? Default value is FALSE.}

\item{weibull_scale}{scalar. Scale parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 0.135. Users may specify a vector of two
values where the first value is used to estimate the weight of the
historical treatment group and the second value is used to estimate the
weight of the historical control group.}

\item{weibull_shape}{scalar. Shape parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 3. Users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group.}

\item{method}{character. Analysis method with respect to estimation of the weight
paramter alpha. Default value "\code{fixed}" estimates alpha once and holds it fixed
throughout the analysis. Alternative method "\code{mc}" estimates alpha for each
Monte Carlo iteration. Currently, only the default method "\code{fixed}" is
supported.}
}
\value{
\code{bdplm} returns an object of class "bdplm".
  The functions \code{summary} and \code{print} are used to obtain and
  print a summary of the results, including user inputs. The \code{plot}
  function displays visual outputs as well.

An object of class "\code{bdplm}" is a list containing at least
the following components:
\describe{
   \itemize{
     \item{\code{alpha_discount}}{
       numeric. Alpha value, the weighting parameter of the historical data.}
     \item{\code{p_hat}}{
       numeric. The posterior probability of the stochastic comparison
       between the current and historical data.}
     \item{\code{posterior_regression}}{
       list. Contains results for the augmented regression analysis. Entries
       are similar to the output of \code{glm} and \code{bayesglm} (from the
        \code{arm} package).}
     \item{\code{posterior_flat_regression}}{
       list. Contains entries similar to \code{posterior_regression} corresponding
       to estimates of the unweighted current data.}
     \item{\code{prior_regression}}{
       to estimates of the historical data.}
   }
}
}
\description{
\code{bdplm} is used to estimate the treatment effect
  in the presence of covariates using the regression analysis
  implementation of the Bayesian discount prior. This function is a barebones
  implementation and does not support \code{summary}, \code{plot}, and
  \code{print} methods.
}
\details{
\code{bdplm} uses a two-stage approach for determining the
  strength of historical data in estimation of an adjusted mean or covariate effect.
  In the first stage, a Weibull distribution function is used as a
  \emph{discount function} that defines the maximum strength of the
  historical data (via \code{weibull_shape}, \code{weibull_scale}, and
  \code{alpha_max}) and discounts based on disagreement with the current data.
  Disagreement between current and historical data is determined by stochastically
  comparing the respective posterior distributions under noninformative priors.
  Here with a two-arm regression analysis, the comparison is the
  proability (\code{p}) that the covariate effect of an historical data indicator is
  significantly different from zero. The comparison metric \code{p} is then
  input into the Weibull discount function and the final strength of the
  historical data is returned (alpha).

  In the second stage, posterior estimation is performed where the discount
  function parameter, \code{alpha}, is used as a fixed value for all posterior
  estimation procedures.

  At minimum, the formula/data must include an intercept and each of historical
  and treatment columns. Any covariates can be included as well. Examplue usage
  that includes the intercept be default could be:
  \code{y ~ hisorical+treatment+baseline}

  In this implementation, current and historical data must be present for both
  treatment and control groups.
}
\examples{
# Simulate regression data
set.seed(3408)
historical <- c(rep(1,50),rep(0,50))
x1         <- c(rnorm(50), rnorm(50))
y          <- c(rnorm(50), rnorm(50)+0.2)

fit1 <- bdpregression(y ~ x1 + historical)
summary(fit1)
print(fit1)
plot(fit1, type="discount")


}
