% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal1arm.R
\name{bdpnormal1arm}
\alias{bdpnormal1arm}
\title{bdpnormal1arm: bdpnormal1arm}
\usage{
bdpnormal1arm(mu_t = 10, sigma2_t = 1, N_t = 10, mu0_t = 10,
  sigma02_t = 1, N0_t = 10, alpha_max = 1, weibull_scale = 0.1,
  weibull_shape = 1, number_mcmc = 10000, two_side = 0,
  inequality = "<", delta = 0)
}
\arguments{
\item{mu_t}{numeric}

\item{sigma2_t}{numeric}

\item{N_t}{numeric}

\item{mu0_t}{numeric}

\item{sigma02_t}{numeric}

\item{N0_t}{numeric}

\item{alpha_max}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{number_mcmc}{numeric}

\item{two_side}{numeric}

\item{inequality}{character}

\item{delta}{numeric}
}
\description{
bdpnormal1arm
}
\details{
bdpnormal1arm
}
\examples{

}

