% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpregression_linear.R
\name{bdpregression_linear}
\alias{bdpregression_linear}
\title{bdpregression_linear: Post Estimate Gaussian}
\usage{
bdpregression_linear(data, formula = y ~ treatment + x, family = "gaussian",
  treatment = "treatment", prior.dist = NULL, prior.mean = 0,
  prior.scale = 1000, prior.df = Inf, mu0 = 1, sigma02 = 0.1,
  weibull_scale = 1, weibull_shape = 1, alpha_max = 1,
  number_mcmc = 10000, two_side = 0)
}
\arguments{
\item{data}{numeric number}

\item{formula}{formula}

\item{family}{family}

\item{treatment}{name of the treatment effect in the data dataframe.}

\item{prior.dist}{prior distribution of the treatment and covariate effects.}

\item{prior.mean}{prior mean of the covariate effects. Default is zero.}

\item{prior.scale}{prior variance of the covariate effects. Default is 1e6.}

\item{prior.df}{prior degrees of freedom of the treatment and covariate effects.}

\item{mu0}{prior mean of the treatment effect}

\item{sigma02}{prior variance of the treatment effect.}

\item{weibull_scale}{weibull_scale}

\item{weibull_shape}{weibull_shape}

\item{alpha_max}{alpha_max}

\item{number_mcmc}{number_mcmc}

\item{two_side}{numeric}
}
\description{
bdpregression_linear
bdpregression_linear
}
\examples{
set.seed(42)
data <- data.frame(y         = rnorm(100, 4, 0.1),
                   x         = c(rnorm(50,1,0.1), rnorm(50,3,0.1)),
                   treatment = c(rep(0,50),rep(1,50)))

fit <- bdpregression_linear(data,
formula       = y ~ treatment + x,
family        = "gaussian",
treatment     = "treatment",
prior.dist    = NULL,
prior.mean    = 0,
prior.scale   = 1000,
prior.df      = Inf,
mu0           = 1,
sigma02       = 0.1,
weibull_scale = 1,
weibull_shape = 1,
alpha_max     = 1,
number_mcmc   = 10000,
two_side      = 0)
### Main parameter of interest:
fit$effect_est
}
