% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helper.R
\name{data_helper}
\alias{data_helper}
\title{data_helper: data_helper}
\usage{
data_helper(current, historical)
}
\arguments{
\item{current}{dataframe}

\item{historical}{dataframe}
}
\description{
data_helper
}
\details{
data_helper
}
\examples{

}

