% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial.R
\name{discount_function_binomial}
\alias{discount_function_binomial}
\alias{discount_function_binomial,ANY-method}
\title{FILL IN}
\usage{
discount_function_binomial(alpha_max = 1, fix_alpha = FALSE,
  number_mcmc = 10000, weibull_scale = 0.135, weibull_shape = 3,
  two_side = TRUE, posterior_flat = NULL, prior = NULL)
}
\description{
FILL IN
}
