% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{final}
\alias{final}
\alias{final,ANY-method}
\title{final}
\usage{
final(posterior_treatment = NULL, posterior_control = NULL)
}
\arguments{
\item{posterior_treatment}{posterior_treatment}

\item{posterior_control}{posterior_control}
}
\description{
final
}
