% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial.R
\name{final_binomial}
\alias{final_binomial}
\alias{final_binomial,ANY-method}
\title{final_binomial}
\usage{
final_binomial(posterior_treatment = NULL, posterior_control = NULL)
}
\arguments{
\item{posterior_treatment}{posterior_treatment}

\item{posterior_control}{posterior_control}
}
\description{
final_binomial
}
