% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{final_normal}
\alias{final_normal}
\alias{final_normal,ANY-method}
\title{final_normal}
\usage{
final_normal(posterior_treatment = NULL, posterior_control = NULL)
}
\arguments{
\item{posterior_treatment}{posterior_treatment}

\item{posterior_control}{posterior_control}
}
\description{
final_normal
}
