% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{mu_post_aug}
\alias{mu_post_aug}
\alias{mu_post_aug,ANY-method}
\title{mu_post_aug}
\usage{
mu_post_aug(mu = NULL, sigma = NULL, N = NULL, mu0 = NULL,
  sigma0 = NULL, N0 = NULL, alpha_discount = NULL, number_mcmc = 10000)
}
\arguments{
\item{mu}{scalar. Mean of the current data.}

\item{sigma}{scalar. Standard deviation of the current data.}

\item{N}{scalar. Number of observations of the current data.}

\item{mu0}{scalar. Mean of the historical data. Required
for single arm (OPC) trials.}

\item{sigma0}{scalar. Standard deviation of the historical treatment
group. Required for single arm (OPC) trials.}

\item{N0}{scalar. Number of observations of the historical treatment
group. Required for single arm (OPC) trials.}

\item{alpha_discount}{scalar. Maximum weight the discount function can apply.
Default is 1. For a two-arm trial, users may specify a vector of two values
where the first value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}

\item{number_mcmc}{scalar. Number of Markov Chain Monte Carlo (MCMC)
simulations. Default is 10000.}

\item{fix_alpha}{logical. Fix alpha at alpha_max? Default value is FALSE.}
}
\description{
mu_post_aug
}
