% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{mu_posterior}
\alias{mu_posterior}
\alias{mu_posterior,ANY-method}
\title{mu_posterior}
\usage{
mu_posterior(mu = NULL, sigma = NULL, N = NULL, mu0 = NULL,
  sigma0 = NULL, N0 = NULL, alpha_max = 1, fix_alpha = FALSE,
  number_mcmc = 10000, weibull_scale = 0.135, weibull_shape = 3,
  two_side = 1)
}
\description{
mu_posterior
}
