% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,bdpbinomial-method}
\alias{plot,bdpbinomial-method}
\title{bdpbinomial Object Plot}
\usage{
\S4method{plot}{bdpbinomial}(x, type = NULL, print = TRUE)
}
\arguments{
\item{x}{object of class \code{bdpbinomial}. The result of a call to the
\code{\link{bdpbinomial}} function.}

\item{type}{character. Optional. Select plot type to print.
Supports the following: "discount" gives the discount function;
"posteriors" gives the posterior plots of the event rates; and
"density" gives the augmented posterior density plot(s).  Leave NULL to
display all plots in sequence.}

\item{print}{logical. Optional. Produce a plot (\code{print = TRUE}; default) or
return a ggplot object (\code{print = FALSE}). When \code{print = FALSE},
it is possible to use \code{ggplot2} syntax to modify the plot appearance.}
}
\description{
\code{plot} method for class \code{bdpbinomial}.
}
\details{
The \code{plot} method for \code{bdpbinomial} returns up to three
  plots: (1) posterior density curves; (2) posterior density of the effect
  of interest; and (3) the discount function. A call to \code{plot} that
  omits the \code{type} input returns one plot at a time and prompts the
  user to click the plot or press return to proceed to the next plot.
  Otherwise, the user can specify a plot \code{type} to display the
  requested plot.
}
