% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,bdpnormal-method}
\alias{plot,bdpnormal-method}
\title{plot}
\usage{
\S4method{plot}{bdpnormal}(x, type = NULL)
}
\arguments{
\item{x}{Result}

\item{type}{character. Optional. Select plot type to print.
Supports the following: "discount" gives the discount function;
"posteriors" gives the posterior plots of the event rates; and
"density" gives the augmented posterior density plot(s).  Leave NULL to
display all plots in sequence.}
}
\description{
normal plot method
}
