% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,bdpregression-method}
\alias{plot,bdpregression-method}
\title{bdpregression Object Plot}
\usage{
\S4method{plot}{bdpregression}(x, type = NULL, print = TRUE)
}
\arguments{
\item{x}{object of class \code{bdpregression}. The result of a call to the
\code{\link{bdpregression}} function.}

\item{type}{character. Optional. Select plot type to print.
Currently supports "discount", the discount function plot.}

\item{print}{logical. Optional. Produce a plot (\code{print = TRUE}; default) or
return a ggplot object (\code{print = FALSE}). When \code{print = FALSE},
it is possible to use \code{ggplot2} syntax to modify the plot appearance.}
}
\description{
\code{plot} method for class \code{bdpregression}.
}
\details{
The \code{plot} method for \code{bdpregression} currently returns
  only the discount function. More plots will be implemented in a future release.
}
