% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,bdpnormal-method}
\alias{plot,bdpbinomial,bdpbinomial-method}
\alias{plot,bdpbinomial-method}
\alias{plot,bdpnormal,bdpnormal-method}
\alias{plot,bdpnormal-method}
\alias{plot,bdpsurvival,bdpsurvival-method}
\alias{plot,bdpsurvival-method}
\title{plot: normal plot method}
\usage{
\S4method{plot}{bdpnormal}(x)

\S4method{plot}{bdpbinomial}(x)

\S4method{plot}{bdpsurvival}(x)
}
\arguments{
\item{x}{Result}
}

