% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,bdpnormal-method}
\alias{plot,bdpnormal-method}
\alias{plot,}
\alias{bdpnormal,}
\alias{bdpbinomial,}
\alias{bdpregression_linear,}
\alias{bdpsurvival}
\title{plot: plot}
\usage{
\S4method{plot}{bdpnormal}(x)
}
\arguments{
\item{x}{Results}
}
\description{
plot
plot
}
