% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial.R
\name{posterior_augment_binomial}
\alias{posterior_augment_binomial}
\alias{posterior_augment_binomial,ANY-method}
\title{FILL IN}
\usage{
posterior_augment_binomial(y = NULL, N = NULL, y0 = NULL, N0 = NULL,
  alpha_max = 1, fix_alpha = FALSE, a0 = 1, b0 = 1,
  number_mcmc = 10000)
}
\description{
FILL IN
}
