% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print,bdpbinomial-method}
\alias{print,bdpbinomial-method}
\title{bdpbinomial Object Print}
\usage{
\S4method{print}{bdpbinomial}(x)
}
\arguments{
\item{x}{object of class \code{bdpbinomial}. The result of a call to the
\code{\link{bdpbinomial}} function.}
}
\description{
\code{print} method for class \code{bdpbinomial}.
}
\details{
Returns same output as a call to
\code{\link[=summary,bdpbinomial-method]{summary}}.
}
