% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print,bdplm-method}
\alias{print,bdplm-method}
\title{bdplm Object Print}
\usage{
\S4method{print}{bdplm}(x)
}
\arguments{
\item{x}{object of class \code{bdplm}. The result of a call to the
\code{\link{bdplm}} function.}
}
\description{
\code{print} method for class \code{bdplm}.
}
\details{
Displays a print of the \code{bdplm} fit and the initial function call.
  The fit includes the estimate of the intercept, treatment effect, and
  covariate effects. The discount function weight estimates are displayed as well.
  If \code{method}="mc", the median estimate of alpha is displayed.
}
