% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print,bdpnormal-method}
\alias{print,bdpnormal-method}
\title{bdpnormal Object Print}
\usage{
\S4method{print}{bdpnormal}(x)
}
\arguments{
\item{x}{object of class \code{bdpnormal}. The result of a call to the
\code{\link{bdpnormal}} function.}
}
\description{
\code{print} method for class \code{bdpnormal}.
}
\details{
Returns same output as a call to \code{\link[=summary,bdpnormal-method]{summary}}.
}
