% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,bdpregression-method}
\alias{print,bdpregression-method}
\title{bdpregression Object Print}
\usage{
\S4method{print}{bdpregression}(x)
}
\arguments{
\item{x}{object of class \code{bdpregression}. The result of a call to the
\code{\link{bdpregression}} function.}
}
\description{
\code{print} method for class \code{bdpregression}.
}
\details{
Returns a print of the model with output similar to \code{\link{lm}}.
}
