% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,bdplm-method}
\alias{summary,bdplm-method}
\title{bdplm Object Summary}
\usage{
\S4method{summary}{bdplm}(object)
}
\arguments{
\item{object}{an object of class \code{bdplm}, a result of a call
to \code{\link{bdplm}}.}
}
\description{
\code{summary} method for class \code{bdplm}.
}
\details{
Displays a summary of the \code{bdplm} fit. Displayed output is similar
  to \code{\link{summary.lm}}. The function call, residuals, and coefficient table
  are displayed. The discount function weight estimates are displayed as well.
  If \code{method}="mc", the median estimate of alpha is displayed.
}
