% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,bdpregression-method}
\alias{summary,bdpregression-method}
\title{bdpregression Object Summary}
\usage{
\S4method{summary}{bdpregression}(object)
}
\arguments{
\item{object}{object of class \code{bdpregression}. The result of a call to the
\code{\link{bdpregression}} function.}
}
\description{
\code{summary} method for class \code{bdpregression}.
}
\details{
Displays a summary of the \code{bdpregression} fit including the
  input data, the stochastic comparison between current and historical
  data, and the resulting historical data weight (alpha). If historical
  data is missing then no stochastic comparison nor weight are displayed.

  ...
}
