% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,bdpsurvival-method}
\alias{summary,bdpsurvival-method}
\title{bdpsurvival Object Summary}
\usage{
\S4method{summary}{bdpsurvival}(object)
}
\arguments{
\item{object}{an object of class \code{bdpsurvival}, a result of a call
to \code{\link{bdpsurvival}}.}
}
\description{
\code{summary} method for class \code{bdpsurvival}.
}
\details{
Displays a summary of the \code{bdpsurvival} fit. The output
  is different, conditional on a one- or two-arm survival analysis.

  In the case of a one-arm analysis, the stochastic comparison between
  current and historical data and the resulting historical data weight
  (alpha) are displayed, followed by a survival table containing
  augmented posterior estimates of the survival probability at each
  time point for the current data.

  When a control arm is present, a two-arm analysis is carried out.
  A two-arm survival analysis is similar to a cox proportional
  hazards analysis, and the displayed summary reflects this. First,
  the stochastic comparison between current and historical data and
  the resulting historical data weight (alpha) are displayed, when
  historical data is present for the respective arm. The displayed
  \code{coef} value is the log-hazard between the augmented treatment
  and control arms (log(treatment) - log(control)). The lower and upper
  95 percent interval limits correspond to the \code{coef} value.
}
