% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,bdpnormal-method}
\alias{summary,bdpbinomial}
\alias{summary,bdpbinomial-method}
\alias{summary,bdpnormal}
\alias{summary,bdpnormal-method}
\alias{summary,bdpsurvival}
\alias{summary,bdpsurvival-method}
\title{summary}
\usage{
\S4method{summary}{bdpnormal}(object)

\S4method{summary}{bdpbinomial}(object)

\S4method{summary}{bdpsurvival}(object)
}
\arguments{
\item{object}{Result}
}
\description{
normal summary method

binomial summary method

survival summary method
}

