% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,bdpnormal-method}
\alias{bdpbinomial,}
\alias{bdpnormal,}
\alias{bdpregression_linear,}
\alias{bdpsurvival}
\alias{summary,}
\alias{summary,bdpnormal-method}
\title{summary: summary}
\usage{
\S4method{summary}{bdpnormal}(object)
}
\arguments{
\item{object}{Results}
}
\description{
summary
summary
}

