% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_discount.R
\name{alpha_discount}
\alias{alpha_discount}
\alias{alpha_discount,ANY-method}
\title{INSERT TEXT HERE}
\usage{
alpha_discount(p_hat = NULL, alpha_max = 1, weibull_scale = 0.135,
  weibull_shape = 3)
}
\arguments{
\item{p_hat}{INSERT TEXT HERE}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. For a two-arm trial, users may specify a vector of two values
where the first value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}

\item{weibull_scale}{scalar. Scale parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 0.135. For a two-arm trial, users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group.}

\item{weibull_shape}{scalar. Shape parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 3. For a two-arm trial, users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group.}
}
\value{
\code{alpha_discount} returns an object of class "alpha_discount".

An object of class \code{alpha_discount} is a list containing at least
the following components:
\describe{
 \item{\code{posterior_treatment}}{
   list. Entries contain values related to the data:}
   \itemize{
    INSERT TEXT HERE
  }
}
}
\description{
\code{alpha_discount} INSERT TEXT HERE
}
\details{
\code{alpha_discount} INSERT TEXT HERE

  For more details, see the \code{alpha_discount} vignette: \cr
  \code{vignette("alpha_discount-vignette", package="bayesDP")}
}
\examples{
# INSERT EXAMPLE HERE

}
\references{
Haddad, T., Himes, A., Thompson, L., Irony, T., Nair, R. MDIC Computer
  Modeling and Simulation working group.(2017) Incorporation of stochastic
  engineering models as prior information in Bayesian medical device trials.
  \emph{Journal of Biopharmaceutical Statistics}, 1-15.
}

