% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLossFunctionNormalRCT.R
\name{BayesianLossFunctionNormalRTC}
\alias{BayesianLossFunctionNormalRTC}
\title{BayesianLossFunctionNormalRTC: BayesianLossFunctionNormalRTC}
\usage{
BayesianLossFunctionNormalRTC(mu, sigma2, N, mu0, sigma02, N0, N0_max,
  weibull_scale, weibull_shape, number_mcmc, H0, two_side, inequality, N0_t,
  N0_c, delta)
}
\arguments{
\item{mu}{numeric}

\item{sigma2}{numeric}

\item{N}{numeric}

\item{mu0}{numeric}

\item{sigma02}{numeric}

\item{N0}{numeric}

\item{N0_max}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{number_mcmc}{numeric}

\item{H0}{numeric}

\item{two_side}{character}

\item{inequality}{character}

\item{N0_t}{numeric}

\item{N0_c}{numeric}

\item{delta}{numeric}
}
\description{
BayesianLossFunctionNormalRTC
}
\details{
BayesianLossFunctionNormalRTC
}
\examples{

}

