% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial.R
\name{bdpbinomial}
\alias{bdpbinomial}
\title{bdpbinomial: bdpbinomial}
\usage{
bdpbinomial(type = c("1arm", "2arm"), ...)
}
\arguments{
\item{type}{numeric}

\item{y_t}{numeric}

\item{N_t}{numeric}

\item{y_c}{numeric}

\item{N_c}{numeric}

\item{y0_t}{numeric}

\item{N0_t}{numeric}

\item{y0_c}{numeric}

\item{N0_c}{numeric}

\item{alpha_max}{numeric}

\item{a0}{numeric}

\item{b0}{numeric}

\item{number_mcmc}{numeric}

\item{weibull_shape}{numeric}

\item{weibull_scale}{numeric}

\item{two_side}{character}

\item{inequality}{character}

\item{delta}{character}
}
\description{
bdpbinomial
}
\details{
bdpbinomial
}
\examples{

}

