% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial1arm.R
\name{bdpbinomial1arm}
\alias{bdpbinomial1arm}
\title{bdpbinomial1arm: bdpbinomial1arm}
\usage{
bdpbinomial1arm(y_t = NULL, N_t = NULL, y0_t = NULL, N0_t = NULL,
  alpha_max = 1, a0 = 1, b0 = 1, number_mcmc = 10000,
  weibull_scale = 0.135, weibull_shape = 3, two_side = 1)
}
\arguments{
\item{y_t}{numeric}

\item{N_t}{numeric}

\item{y0_t}{numeric}

\item{N0_t}{numeric}

\item{alpha_max}{numeric}

\item{a0}{numeric}

\item{b0}{numeric}

\item{number_mcmc}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{two_side}{character}
}
\description{
This code is used for estimating posterior samples from a binary outcome
where an informative prior is used. The prior weight is determined using a
discount function. In addition this code simulate many trials in order to get
trial characteristics you must specify the parameters of the discount function
as well as the maximum strength for the prior. This code assumes a
non-adaptive trial.
This code is modeled after the methodologies developed by the MDIC working
group: "Informing clinical trials using bench & simulations"
Developer: Tarek Haddad
Tarek.D.Haddad@Medtronic.com
Last modified:1/26/2016
}
\details{
bdpbinomial1arm
}
\examples{

}

