% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial2arm.R
\name{bdpbinomial2arm}
\alias{bdpbinomial2arm}
\title{bdpbinomial2arm: bdpbinomial2arm}
\usage{
bdpbinomial2arm(y_t = 1, N_t = 400, y_c = 5, N_c = 400, y0_t = 1,
  N0_t = 200, y0_c = 5, N0_c = 200, alpha_max = 1, a0 = 1, b0 = 1,
  number_mcmc = 10000, weibull_scale = 0.05, weibull_shape = 2,
  two_side = 0, inequality = "<", delta = 0)
}
\arguments{
\item{y_t}{numeric}

\item{N_t}{numeric}

\item{y_c}{numeric}

\item{N_c}{numeric}

\item{y0_t}{numeric}

\item{N0_t}{numeric}

\item{y0_c}{numeric}

\item{N0_c}{numeric}

\item{alpha_max}{numeric}

\item{a0}{numeric}

\item{b0}{numeric}

\item{number_mcmc}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{two_side}{character}

\item{inequality}{character}

\item{delta}{character}
}
\description{
bdpbinomial2arm
}
\details{
bdpbinomial2arm
}
\examples{

}

