% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdplm.R
\name{bdplm}
\alias{bdplm}
\alias{bdplm,ANY-method}
\title{Bayesian Discount Prior: Two-Arm Linear Regression}
\usage{
bdplm(formula = formula, data = data, prior_mean = 0, prior_sigma = 1,
  number_mcmc_alpha = 10000, number_mcmc_sigmagrid = 1000,
  number_mcmc_sigma = 100, number_mcmc_beta = 10000, alpha_max = 1,
  fix_alpha = FALSE, weibull_scale = 0.135, weibull_shape = 3,
  method = "fixed")
}
\arguments{
\item{formula}{an object of class "formula." See "Details" for
more information, including specfication of historical and treatment
data indicators.}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)
containing the variables in the model. If not found in data,
the variables are taken from environment(formula), typically
the environment from which bdplm is called.}

\item{prior_mean}{historical mean for the coefficients, specified as a
a vector. The length of \code{prior_mean} must be equal to the number
of adjusting covariates + 2 (ignore the intercept in this case). See
"Details" for more information.}

\item{prior_sigma}{prior standard deviation for the coefficient, specified
as a vector. The length of \code{prior_sigma} must be equal to the number
of adjusting covariates + 2 (ignore the intercept in this case). See
"Details" for more information.}

\item{number_mcmc_alpha}{scalar. Number of Monte Carlo
simulations for estimating the historical data weight. Default is 10000.}

\item{number_mcmc_sigmagrid}{scalar. Grid size for computing sigma.
Default is 1000. See "Details" for more information.}

\item{number_mcmc_sigma}{scalar. Number of Monte Carlo simulations for
estimating sigma. Default is 1000. See "Details" for more information.}

\item{number_mcmc_beta}{scalar. Number of Monte Carlo simulations for
estimating beta, the vector of regression coefficients. Default is 10000.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. Users may specify a vector of two values where the first
value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}

\item{fix_alpha}{logical. Fix alpha at alpha_max? Default value is FALSE.}

\item{weibull_scale}{scalar. Scale parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 0.135. Users may specify a vector of two
values where the first value is used to estimate the weight of the
historical treatment group and the second value is used to estimate the
weight of the historical control group.}

\item{weibull_shape}{scalar. Shape parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 3. Users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group.}

\item{method}{character. Analysis method with respect to estimation of the
weight paramter alpha. Default value "\code{fixed}" estimates alpha once
and holds it fixed throughout the analysis. Alternative method
"\code{mc}" estimates alpha for each Monte Carlo iteration. Currently, only
the default method "\code{fixed}" is supported.}
}
\value{
\code{bdplm} returns an object of class "bdplm".

An object of class "\code{bdplm}" is a list containing at least
the following components:
\describe{
   \itemize{
     \item{\code{posterior}}{
       data frame. The posterior draws of the covariates, the intercept, and
       the treatment effect. The grid of sigma values are included.}
     \item{\code{alpha_discount}}{
       numeric. The posterior probability of the stochastic comparison
       between the current and historical data.}
   }
}
}
\description{
\code{bdplm} is used to estimate the treatment effect
  in the presence of covariates using the regression analysis
  implementation of the Bayesian discount prior. This function is a
  barebones implementation and does not support \code{summary},
  \code{plot}, and \code{print} methods.
}
\details{
\code{bdplm} uses a two-stage approach for determining the
  strength of historical data in estimation of an adjusted mean or covariate
  effect. In the first stage, a Weibull distribution function is used as a
  \emph{discount function} that defines the maximum strength of the
  historical data (via \code{weibull_shape}, \code{weibull_scale}, and
  \code{alpha_max}) and discounts based on disagreement with the current data.
  Disagreement between current and historical data is determined by stochastically
  comparing the respective posterior distributions under noninformative priors.
  Here with a two-arm regression analysis, the comparison is the
  proability (\code{p}) that the covariate effect of an historical data indicator is
  significantly different from zero. The comparison metric \code{p} is then
  input into the Weibull discount function and the final strength of the
  historical data is returned (alpha).

  In the second stage, posterior estimation is performed where the discount
  function parameter, \code{alpha}, is used as a fixed value for all posterior
  estimation procedures.

  At minimum, the formula/data must include an intercept and each of historical
  and treatment columns. Any covariates can be included as well. Example usage
  that includes the intercept by default could be:
  \code{y ~ hisorical+treatment+baseline}.

  In this implementation, current and historical data must be present for both
  treatment and control groups. The covariate prior parameters \code{prior_mean}
  and \code{prior_sigma} must be input as a vector with elements in the order
  given by the formula.
}
\examples{
# Simulate  data
n_t  <- 100
n_c  <- 100
n_t0 <- 250
n_c0 <- 250
x          <- rnorm(n_t+n_c+n_t0+n_c0, 34, 5)
treatment  <- c(rep(1, n_t+n_t0), rep(0, n_c+n_c0))
historical <- c(rep(0, n_t), rep(1,n_t0), rep(0, n_c), rep(1,n_c0))

Y <- treatment + 1000*historical + x*3.5 + rnorm(n_t+n_c+n_t0+n_c0)
df <- data.frame(Y=Y, treatment=treatment, historical=historical, x=x)

fit <- bdplm(Y ~ treatment+historical+x, data=df,
             prior_mean  = rep(0,3),
             prior_sigma = rep(1e4,3))


}

