% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{bdpnormal}
\alias{bdpnormal}
\title{bdpnormal: bdpnormal}
\usage{
bdpnormal(type = c("1arm", "2arm"), ...)
}
\arguments{
\item{type}{character}

\item{mu_t}{numeric}

\item{sigma2_t}{numeric}

\item{N_t}{numeric}

\item{mu_c}{numeric}

\item{sigma2_c}{numeric}

\item{N_c}{numeric}

\item{mu0_t}{numeric}

\item{sigma02_t}{numeric}

\item{N0_t}{numeric}

\item{mu0_c}{numeric}

\item{sigma02_c}{numeric}

\item{N0_c}{numeric}

\item{alpha_max}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{number_mcmc}{numeric}

\item{two_side}{character}

\item{inequality}{character}

\item{delta}{numeric}
}
\description{
bdpnormal
}
\details{
bdpnormal
}
\examples{

}

