% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{bdpnormal}
\alias{bdpnormal}
\title{Bayesian Discount Prior: Gaussian mean values}
\usage{
bdpnormal(mu_t = NULL, sigma_t = NULL, N_t = NULL, mu0_t = NULL,
  sigma0_t = NULL, N0_t = NULL, mu_c = NULL, sigma_c = NULL,
  N_c = NULL, mu0_c = NULL, sigma0_c = NULL, N0_c = NULL,
  alpha_max = 1, weibull_scale = 0.135, weibull_shape = 3,
  number_mcmc = 10000, two_side = 1)
}
\arguments{
\item{mu_t}{scalar. The mean of the current treatment group.}

\item{sigma_t}{scalar. The standard deviation of the current treatment
group.}

\item{N_t}{scalar. The number of observations of the current treatment
group.}

\item{mu0_t}{scalar. The mean of the historical treatment group. Required
for \code{type="1arm"} (OPC trials).}

\item{sigma0_t}{scalar. The standard deviation of the historical treatment
group. Required for \code{type="1arm"} (OPC trials).}

\item{N0_t}{scalar. The number of observations of the historical treatment
group. Required for \code{type="1arm"} (OPC trials).}

\item{mu_c}{scalar. The mean of the current control group. Required for
\code{type="2arm"} (RCT trials).}

\item{sigma_c}{scalar. The standard deviation of the current control group.
Required for \code{type="2arm"} (RCT trials).}

\item{N_c}{scalar. The number of observations of the current control group.
Required for \code{type="2arm"} (RCT trials).}

\item{mu0_c}{scalar. The mean of the historical control group.}

\item{sigma0_c}{scalar. The standard deviation of the historical control
group.}

\item{N0_c}{scalar. The number of observations of the historical control
group.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. For \code{type="2arm"}, users may optionally specify
a vector of two values where the first value is used to weight the
historical treatment group and the second value is used to weight
the historical control group.}

\item{weibull_scale}{scalar. Scale parameter of the Weibull discount
function used to compute alpha, the weight parameter of the historical
data. Default value is 0.135. Two values have special treatment:
\code{0} and \code{Inf}. For \code{weibull_scale = 0}, alpha is set
to 0, i.e., no weight. For \code{weibull_scale = Inf}, alpha is set
to 1, i.e., full weight. For \code{type="2arm"}, users may optionally
specify a vector of two values where the first value is used to
estimate the weight of the historical treatment group and the second
value is used to estimate the weight of the historical control group.}

\item{weibull_shape}{scalar. Shape parameter of the Weibull discount
function used to compute alpha, the weight parameter of the historical
data. Default value is 3. For \code{type="2arm"}, users may optionally
specify a vector of two values where the first value is used to
estimate the weight of the historical treatment group and the second
value is used to estimate the weight of the historical control group.}

\item{number_mcmc}{scalar. Number of Markov Chain Monte Carlo (MCMC)
simulations. Default is 1e4.}

\item{two_side}{scalar. Indicator of two-sided test for the discount
function. Default value is 1.}
}
\value{
\code{bdpnormal} returns an object of class "bdpnormal".
The functions \code{summary} and \code{print} are used to obtain and
print a summary of the results, including user inputs. The \code{plot}
function displays visual outputs as well.
An object of class "\code{bdpnormal} " is a list containing at least
the following components:
}
\description{
The pdpnormal function is used for estimating posterior samples from a
Gaussian outcome where an informative prior is used. The prior weight
is determined using a discount function. This code is modeled after
the methodologies developed by the MDIC working group: "Informing
clinical trials using bench & simulations."
}
\details{
Many, many, many details to come. In fact, the best details. Believe
me, I know a thing or two about building details.
}
\examples{
# One-arm trial (OPC) example
s
s
s
s
}

