% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal2arm.R
\name{bdpnormal2arm}
\alias{bdpnormal2arm}
\title{bdpnormal2arm: bdpnormal2arm}
\usage{
bdpnormal2arm(mu_t = 10, sigma2_t = 2, N_t = 20, mu_c = 15,
  sigma2_c = 2, N_c = 20, mu0_t = 12, sigma02_t = 1, N0_t = 10,
  mu0_c = 12, sigma02_c = 1, N0_c = 10, alpha_max = 1,
  weibull_scale = 0.2, weibull_shape = 2, number_mcmc = 10000,
  two_side = 1, inequality = "<", delta = 0)
}
\arguments{
\item{mu_t}{numeric}

\item{sigma2_t}{numeric}

\item{N_t}{numeric}

\item{mu_c}{numeric}

\item{sigma2_c}{numeric}

\item{N_c}{numeric}

\item{mu0_t}{numeric}

\item{sigma02_t}{numeric}

\item{N0_t}{numeric}

\item{mu0_c}{numeric}

\item{sigma02_c}{numeric}

\item{N0_c}{numeric}

\item{alpha_max}{numeric}

\item{weibull_scale}{numeric}

\item{weibull_shape}{numeric}

\item{number_mcmc}{numeric}

\item{two_side}{character}

\item{inequality}{character}

\item{delta}{numeric}
}
\description{
This code is used for estimating posterior samples from a
Gaussian outcome where an informative prior is used. The prior weight is
determined using a loss function. In addition this code simulate many
trials in order to get trial characteristics you must specify the
parameters of the loss function as well as the maximum strength for the
prior. This code assumes a non-adaptive trial. This code is modeled after
the methodologies developed by the MDIC working group:
'Informing clinical trials using bench & simulations'
Developer: Tarek Haddad
Tarek.D.Haddad@Medtronic.com
Last modified: 9/10/2016
}
\details{
bdpnormal2arm
}
\examples{

}

