% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpsurvival.R
\name{bdpsurvival}
\alias{bdpsurvival}
\alias{bdpsurvival,ANY-method}
\title{Bayesian Discount Prior: Survival Analysis}
\usage{
bdpsurvival(formula = formula, data = data, breaks = NULL, a0 = 0.1,
  b0 = 0.1, surv_time = NULL, alpha_max = 1, number_mcmc = 10000,
  weibull_scale = 0.135, weibull_shape = 3, two_side = 1)
}
\arguments{
\item{formula}{an object of class "formula." Must have a survival object on
the left side and exactly two inputs on the right side: treatment and
historical. See "Details" for more information.}

\item{data}{data frame. A data frame with columns 'time', 'status',
'treatment', and historical.' See "Details" for required structure.}

\item{breaks}{vector. Breaks (intervals) used to compose the breaks of the
piecewise exponential model.}

\item{a0}{scalar. Prior value for the gamma shape. Default is 1.}

\item{b0}{scalar. Prior value for the gamma rate. Default is 1.}

\item{surv_time}{scalar. Survival time of interest for computing the the
probability of survival for a single arm, i.e., an OPC trial. Default is
median survival time.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. For a two-arm trial, users may specify a vector of two values
where the first value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}

\item{number_mcmc}{scalar. Number of Markov Chain Monte Carlo (MCMC)
simulations. Default is 10000.}

\item{weibull_scale}{scalar. Scale parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 0.135. Two values have special treatment: 0 and Inf. For
weibull_scale = 0, alpha is set to 0, i.e., no weight. For
weibull_scale = Inf, alpha is set to 1, i.e., full weight. For a two-arm
trial, users may specify a vector of two values where the first value is
used to estimate the weight of the historical treatment group and the
second value is used to estimate the weight of the historical control
group.}

\item{weibull_shape}{scalar. Shape parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 3. For a two-arm trial, users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group.}

\item{two_side}{scalar. Indicator of two-sided test for the discount
function. Default value is 1.}
}
\value{
\code{bdpsurvival} returns an object of class "bdpsurvival".
The functions \code{summary} and \code{print} are used to obtain and
print a summary of the results, including user inputs. The \code{plot}
function displays visual outputs as well.
An object of class "\code{bdpsurvival}" is a list containing at least
the following components:
}
\description{
\code{bdpsurvival} is used to estimate the survival probability
  (single arm trial; OPC) or hazard ratio (two-arm trial; RCT) using the
  survival analysis implementation of the Bayesian discount prior.
}
\details{
Many, many, many details to come. In fact, the best details. Believe
me, I know a thing or two about building details.
}
\examples{
# One-arm trial (OPC) example
# Simulate survival data for a single arm (OPC) trial
time   <- c(rexp(50, rate=1/20), rexp(50, rate=1/10))
status <- c(rexp(50, rate=1/30), rexp(50, rate=1/30))
status <- ifelse(time < status, 1, 0)

# Collect data into a dataframe
example_surv_1arm <- data.frame(status     = status,
                                time       = time,
                                historical = c(rep(1,50),rep(0,50)),
                                treatment  = 1)

#fitSurv <- bdpsurvival(Surv(time, status) ~ historical + treatment,
#                       data = example_surv_1arm)


}
