% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{Discount_function}
\alias{Discount_function}
\alias{Discount_function,ANY-method}
\title{Discount_function}
\usage{
Discount_function(mu = NULL, sigma = NULL, N = NULL, mu0 = NULL,
  sigma0 = NULL, N0 = NULL, alpha_max = 1, fix_alpha = FALSE,
  number_mcmc = 10000, weibull_scale = 0.135, weibull_shape = 3,
  two_side = 1)
}
\arguments{
\item{mu}{scalar. Mean of the current data.}

\item{sigma}{scalar. Standard deviation of the current data.}

\item{N}{scalar. Number of observations of the current data.}

\item{mu0}{scalar. Mean of the historical data. Required
for single arm (OPC) trials.}

\item{sigma0}{scalar. Standard deviation of the historical treatment
group. Required for single arm (OPC) trials.}

\item{N0}{scalar. Number of observations of the historical treatment
group. Required for single arm (OPC) trials.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. For a two-arm trial, users may specify a vector of two values
where the first value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}

\item{fix_alpha}{logical. Fix alpha at alpha_max? Default value is FALSE.}

\item{number_mcmc}{scalar. Number of Markov Chain Monte Carlo (MCMC)
simulations. Default is 10000.}

\item{weibull_scale}{scalar. Scale parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 0.135. For a two-arm trial, users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group.}

\item{weibull_shape}{scalar. Shape parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 3. For a two-arm trial, users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group.}

\item{two_side}{scalar. Indicator of two-sided test for the discount
function. Default value is 1.}
}
\description{
Discount_function
}
