% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpnormal.R
\name{mu_post_aug}
\alias{mu_post_aug}
\alias{mu_post_aug,ANY-method}
\title{FILL IN}
\usage{
mu_post_aug(mu = NULL, sigma = NULL, N = NULL, mu0 = NULL,
  sigma0 = NULL, N0 = NULL, alpha_discount = NULL, number_mcmc = 10000)
}
\description{
FILL IN
}
