% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,bdpbinomial-method}
\alias{plot,bdpbinomial-method}
\title{plot}
\usage{
\S4method{plot}{bdpbinomial}(x, type = NULL)
}
\arguments{
\item{x}{Result}

\item{type}{character. Optional. Select plot type to print.
Supports "discount" for the discount function. Leave NULL to
display all plots in sequence.}
}
\description{
binomial plot method
}
