% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,bdpsurvival-method}
\alias{plot,bdpsurvival-method}
\title{plot}
\usage{
\S4method{plot}{bdpsurvival}(x, type = NULL)
}
\arguments{
\item{x}{Result}

\item{type}{character. Optional. Select plot type to print.
Supports the following: "discount" gives the discount function and
"survival" gives the survival curves.  Leave NULL to
display all plots in sequence.}
}
\description{
survival plot method
}
