% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial1arm.R, R/bdpbinomial2arm.R, R/bdpnormal.R, R/bdpregression_linear.R
\docType{methods}
\name{plot,bdpbinomial1arm-method}
\alias{plot,bdpbinomial1arm-method}
\alias{plot,bdpbinomial2arm-method}
\alias{plot,bdpnormal-method}
\alias{plot,bdpregression_linear-method}
\title{plot: plot}
\usage{
\S4method{plot}{bdpbinomial1arm}(x)

\S4method{plot}{bdpbinomial2arm}(x)

\S4method{plot}{bdpnormal}(x)

\S4method{plot}{bdpregression_linear}(x)
}
\arguments{
\item{x}{bdpbinomial1arm}

\item{x}{bdpbinomial2arm}

\item{x}{bdpnormal}

\item{x}{bdpregression_linear}
}
\description{
plot

plot

plot

plot
}
\details{
plot

plot

plot

plot
}
\examples{




}

