% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial1arm.R, R/bdpbinomial2arm.R, R/bdpnormal1arm.R, R/bdpnormal2arm.R
\docType{methods}
\name{plot,bdpbinomial1arm-method}
\alias{plot,bdpbinomial1arm-method}
\alias{plot,bdpbinomial2arm-method}
\alias{plot,bdpnormal1arm-method}
\alias{plot,bdpnormal2arm-method}
\title{plot: plot}
\usage{
\S4method{plot}{bdpbinomial1arm}(x)

\S4method{plot}{bdpbinomial2arm}(x)

\S4method{plot}{bdpnormal1arm}(x)

\S4method{plot}{bdpnormal2arm}(x)
}
\arguments{
\item{x}{bdpbinomial1arm}

\item{x}{bdpbinomial2arm}

\item{x}{bdpnormal1arm}

\item{x}{bdpnormal2arm}
}
\description{
plot

plot

plot

plot
}
\details{
plot

plot

plot

plot
}
\examples{




}

