% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,bdpnormal-method}
\alias{plot,bdpbinomial-method}
\alias{plot,bdpnormal-method}
\alias{plot,bdpregression_linear-method}
\alias{plot,bdpsurvival-method}
\title{plot: plot}
\usage{
\S4method{plot}{bdpnormal}(x)

\S4method{plot}{bdpbinomial}(x)

\S4method{plot}{bdpregression_linear}(x)

\S4method{plot}{bdpsurvival}(x)
}
\arguments{
\item{x}{bdpnormal}

\item{x}{bdpbinomial}

\item{x}{bdpregression_linear}

\item{x}{bdpsurvival}
}
\description{
plot
plot

plot
plot

plot
plot

plot
plot
}

