% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_est_gaussian.R
\docType{methods}
\name{post_est_gaussian}
\alias{post_est_gaussian}
\alias{post_est_gaussian,data.frame,formula,numeric,numeric,numeric,numeric,numeric,missing,numeric-method}
\title{post_est_gaussian: Post Estimate Gaussian}
\usage{
post_est_gaussian(data, formula, prior_mu, prior_sigma, weibull_scale,
  weibull_shape, alpha_max, N0_max = NULL, N_mcmc)


  \S4method{post_est_gaussian}{data.frame,formula,numeric,numeric,numeric,numeric,numeric,missing,numeric}(data,
  formula, prior_mu, prior_sigma, weibull_scale, weibull_shape, alpha_max,
  N0_max = NULL, N_mcmc)
}
\arguments{
\item{data}{numeric number}

\item{formula}{formula}

\item{prior_mu}{prior_mu}

\item{prior_sigma}{prior_sigma}

\item{weibull_scale}{weibull_scale}

\item{weibull_shape}{weibull_shape}

\item{alpha_max}{alpha_max}

\item{N0_max}{N0_max}

\item{N_mcmc}{N_mcmc}
}
\description{
Post Estimate Gaussian
}
\details{
Post Estimate Gaussian
}
\examples{
set.seed(42)
data <- data.frame(y         = rnorm(100, 4, 0.1),
                   x         = c(rnorm(50,1,0.1), rnorm(50,3,0.1)),
                   treatment = c(rep(0,50),rep(1,50)))

fit <- post_est_gaussian(data          = data,
                         formula       = y ~ treatment + x,
                         prior_mu      = 1,
                         prior_sigma   = 0.1,
                         weibull_scale = 1,
                         weibull_shape = 1,
                         alpha_max     = 1,
                         N_mcmc        = 10000)

### Main parameter of interest:
fit$effect_est

}

