% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial.R
\name{posterior_augment_binomial}
\alias{posterior_augment_binomial}
\alias{posterior_augment_binomial,ANY-method}
\title{posterior_augment_binomial}
\usage{
posterior_augment_binomial(y = NULL, N = NULL, y0 = NULL, N0 = NULL,
  alpha_discount = NULL, fix_alpha = FALSE, a0 = 1, b0 = 1,
  number_mcmc = 10000)
}
\arguments{
\item{y}{scalar. Number of events for the current data.}

\item{N}{scalar. Sample size of the current data.}

\item{y0}{scalar. Number of events for the historical data.}

\item{N0}{scalar. Sample size of the historical data.}

\item{fix_alpha}{logical. Fix alpha at alpha_max? Default value is FALSE.}

\item{a0}{scalar. Prior value for the beta rate. Default is 1.}

\item{b0}{scalar. Prior value for the beta rate. Default is 1.}

\item{number_mcmc}{scalar. Number of Markov Chain Monte Carlo (MCMC)
simulations. Default is 10000.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. For a two-arm trial, users may specify a vector of two values
where the first value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}
}
\description{
posterior_augment_binomial
}
