% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppexp.R
\name{ppexp}
\alias{ppexp}
\title{Compute cdf of the piecewise exponential distribution}
\usage{
ppexp(q, x, cuts)
}
\arguments{
\item{q}{scalar. The time point at which the cdf is to be computed.}

\item{x}{vector of hazard rates.}

\item{cuts}{vector of the same length as x, giving the times at which the
rate changes, i.e., the interval cut points. The first element of cuts
should be 0, and cuts should be in increasing order.}
}
\value{
The cumulative distribution function computed at time \code{q},
hazard(s) \code{x}, and cut points \code{cuts}.
}
\description{
\code{ppexp} is used to compute the cumulative distribution
function of the piecewise exponential distribution. The function is ported from
R to C++ via code adapted from the \code{msm} package.
}
\details{
The ppexp function is written in C++ and adapted from R code used
in the \code{msm} package.
}
\examples{
# Single vector of hazard rates. Returns a single cdf value.
q    <- 12
x    <- c(0.25,0.3,0.35,0.4)
cuts <- c(0,6,12,18)
pp   <- ppexp(q,x,cuts)

# Matrix of multiple vectors of hazard rates. Returns 10 cdf values.
x  <- matrix(rgamma(4*10, 0.1, 0.1), nrow=10)
pp <- ppexp(q,x,cuts)
}

