% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,bdpnormal-method}
\alias{print,bdpnormal-method}
\alias{print,bdpnormal,bdpnormal-method}
\alias{print,bdpbinomial-method}
\alias{print,bdpbinomial,bdpbinomial-method}
\alias{print,bdpregression_linear-method}
\alias{print,bdpregression_linear,bdpregression_linear-method}
\alias{print,bdpsurvival-method}
\alias{print,bdpsurvival,bdpsurvival-method}
\title{print: print}
\usage{
\S4method{print}{bdpnormal}(x)

\S4method{print}{bdpbinomial}(x)

\S4method{print}{bdpregression_linear}(x)

\S4method{print}{bdpsurvival}(x)
}
\arguments{
\item{x}{Result}
}
\description{
print
print

print
print

print
print

print
print
}
