% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial2arm.R, R/bdpnormal1arm.R, R/bdpnormal2arm.R
\docType{methods}
\name{print,bdpbinomial2arm-method}
\alias{print,bdpbinomial2arm-method}
\alias{print,bdpnormal1arm-method}
\alias{print,bdpnormal2arm-method}
\title{print: print}
\usage{
\S4method{print}{bdpbinomial2arm}(x)

\S4method{print}{bdpnormal1arm}(x)

\S4method{print}{bdpnormal2arm}(x)
}
\arguments{
\item{x}{bdpbinomial2arm}

\item{x}{bdpnormal1arm}

\item{x}{bdpnormal2arm}
}
\description{
print

print

print
}
\details{
print

print

print
}
\examples{



}

