% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial.R, R/bdpnormal.R,
%   R/bdpregression_linear.R
\docType{methods}
\name{print,bdpbinomial-method}
\alias{print,bdpbinomial-method}
\alias{print,bdpnormal-method}
\alias{print,bdpregression_linear-method}
\title{print: print}
\usage{
\S4method{print}{bdpbinomial}(x)

\S4method{print}{bdpnormal}(x)

\S4method{print}{bdpregression_linear}(x)
}
\arguments{
\item{x}{bdpbinomial}

\item{x}{bdpnormal}

\item{x}{bdpregression_linear}
}
\description{
print

print

print
}
\details{
print

print

print
}
\examples{



}
