% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,bdpnormal-method}
\alias{print,bdpnormal-method}
\alias{print,}
\alias{bdpnormal,}
\alias{bdpbinomial,}
\alias{bdpregression_linear,}
\alias{bdpsurvival}
\title{print: print}
\usage{
\S4method{print}{bdpnormal}(x)
}
\arguments{
\item{x}{Results}
}
\description{
print
print
}
