% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_discount.R
\name{probability_discount}
\alias{probability_discount}
\title{INSERT TEXT HERE}
\usage{
probability_discount(mu = NULL, sigma = NULL, N = NULL, mu0 = NULL,
  sigma0 = NULL, N0 = NULL, number_mcmc = 10000, two_side = TRUE)
}
\arguments{
\item{mu}{scalar. Mean of the current data.}

\item{sigma}{scalar. Standard deviation of the current data.}

\item{N}{scalar. Number of observations of the current data.}

\item{mu0}{scalar. Mean of the historical data.}

\item{sigma0}{scalar. Standard deviation of the historical data.}

\item{N0}{scalar. Number of observations of the historical data.}

\item{number_mcmc}{scalar. Number of Monte Carlo simulations. Default is 10000.}

\item{two_side}{logical. Indicator of two-sided test for the discount
function. Default value is TRUE.}
}
\value{
\code{probability_discount} returns an object of class "probability_discount".

An object of class \code{probability_discount} is a list containing at least
the following components:
\describe{
 \item{\code{posterior_treatment}}{
   list. Entries contain values related to the data:}
   \itemize{
    INSERT TEXT HERE
  }
}
}
\description{
\code{probability_discount} INSERT TEXT HERE
}
\details{
\code{probability_discount} INSERT TEXT HERE

  For more details, see the \code{probability_discount} vignette: \cr
  \code{vignette("probability_discount-vignette", package="bayesDP")}
}
\examples{
# INSERT EXAMPLE HERE

}
\references{
Haddad, T., Himes, A., Thompson, L., Irony, T., Nair, R. MDIC Computer
  Modeling and Simulation working group.(2017) Incorporation of stochastic
  engineering models as prior information in Bayesian medical device trials.
  \emph{Journal of Biopharmaceutical Statistics}, 1-15.
}
\seealso{
\code{\link[=summary,probability_discount-method]{summary}},
  \code{\link[=print,probability_discount-method]{print}},
  and \code{\link[=plot,probability_discount-method]{plot}} for details of each of the
  supported methods.
}

