% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,bdpbinomial-method}
\alias{summary,bdpbinomial-method}
\title{summary}
\usage{
\S4method{summary}{bdpbinomial}(object)
}
\arguments{
\item{object}{Result}
}
\description{
binomial summary method
}
