% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,bdpnormal-method}
\alias{summary,bdpnormal-method}
\title{summary}
\usage{
\S4method{summary}{bdpnormal}(object)
}
\arguments{
\item{object}{Result}
}
\description{
normal summary method
}

