% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,bdpsurvival-method}
\alias{summary,bdpsurvival-method}
\title{summary}
\usage{
\S4method{summary}{bdpsurvival}(object)
}
\arguments{
\item{object}{Result}
}
\description{
survival summary method
}
