% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,bdpnormal-method}
\alias{summary,bdpbinomial-method}
\alias{summary,bdpnormal-method}
\alias{summary,bdpregression_linear-method}
\alias{summary,bdpsurvival-method}
\title{summary: summary}
\usage{
\S4method{summary}{bdpnormal}(object)

\S4method{summary}{bdpbinomial}(object)

\S4method{summary}{bdpregression_linear}(object)

\S4method{summary}{bdpsurvival}(object)
}
\arguments{
\item{object}{bdpnormal}

\item{object}{bdpbinomial}

\item{object}{bdpregression_linear}

\item{object}{bdpsurvival}
}
\description{
summary
summary

summary
summary

summary
summary

summary
summary
}

