% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial1arm.R, R/bdpbinomial2arm.R, R/bdpnormal1arm.R, R/bdpnormal2arm.R
\docType{methods}
\name{summary,bdpbinomial1arm-method}
\alias{summary,bdpbinomial1arm-method}
\alias{summary,bdpbinomial2arm-method}
\alias{summary,bdpnormal1arm-method}
\alias{summary,bdpnormal2arm-method}
\title{summary: summary}
\usage{
\S4method{summary}{bdpbinomial1arm}(object)

\S4method{summary}{bdpbinomial2arm}(object)

\S4method{summary}{bdpnormal1arm}(object)

\S4method{summary}{bdpnormal2arm}(object)
}
\arguments{
\item{object}{bdpbinomial1arm}

\item{object}{bdpbinomial2arm}

\item{object}{bdpnormal1arm}

\item{object}{bdpnormal2arm}
}
\description{
summary

summary

summary

summary
}
\details{
summary

summary

summary

summary
}
\examples{




}

