% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial1arm.R, R/bdpbinomial2arm.R, R/bdpnormal.R
\docType{methods}
\name{summary,bdpbinomial1arm-method}
\alias{summary,bdpbinomial1arm-method}
\alias{summary,bdpbinomial2arm-method}
\alias{summary,bdpnormal-method}
\title{summary: summary}
\usage{
\S4method{summary}{bdpbinomial1arm}(object)

\S4method{summary}{bdpbinomial2arm}(object)

\S4method{summary}{bdpnormal}(object)
}
\arguments{
\item{object}{bdpbinomial1arm}

\item{object}{bdpbinomial2arm}

\item{object}{bdpnormal}
}
\description{
summary

summary

summary
}
\details{
summary

summary

summary
}
\examples{



}

