% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpbinomial.R, R/bdpnormal.R, R/bdpsurvival.R
\docType{methods}
\name{summary,bdpbinomial-method}
\alias{summary,bdpbinomial-method}
\alias{summary,bdpnormal-method}
\alias{summary,bdpsurvival-method}
\title{summary: summary}
\usage{
\S4method{summary}{bdpbinomial}(object)

\S4method{summary}{bdpnormal}(object)

\S4method{summary}{bdpsurvival}(object)
}
\arguments{
\item{object}{bdpbinomial}

\item{object}{bdpnormal}

\item{object}{bdpsurvival}
}
\description{
summary
summary

summary
summary

summary
summary
}

